# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skald AI API",
    "description": "AI services API for speech-to-text transcription, LLM chat completions, and subscription management",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/subscription/delete-account": {
      "delete": {
        "tags": ["subscription"],
        "operationId": "delete_account",
        "responses": {
          "200": {
            "description": "Account deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/support/chatwoot/conversations/{conversation_id}/messages": {
      "get": {
        "tags": ["chatwoot"],
        "operationId": "get_messages",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "source_id",
            "in": "query",
            "description": "Contact source ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Chatwoot API error"
          }
        }
      },
      "post": {
        "tags": ["chatwoot"],
        "operationId": "send_message",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Chatwoot API error"
          }
        }
      }
    },
...
    "/ticket/linear/list-teams": {
      "post": {
        "tags": ["ticket"],
        "operationId": "linear_list_teams",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinearListTeamsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Linear teams fetched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/ticket/linear/list-tickets": {
      "post": {
        "tags": ["ticket"],
        "operationId": "linear_list_tickets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinearListTicketsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Linear tickets fetched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CollectionPage": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionRef"
            }
          },
          "next_cursor": {
            "type": ["string", "null"]
          }
        }
      },
      "CollectionRef": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "type": ["string", "null"],
            "description": "Short identifier. GitHub: \"owner/repo\", Linear: team key (e.g., \"ENG\")."
          },
          "name": {
            "type": "string",
            "description": "Display name. GitHub: \"owner/repo\", Linear: team name."
          },
          "url": {
            "type": ["string", "null"]
          }
        }
      },
...
      "DeleteAccountResponse": {
        "type": "object",
        "required": ["deleted"],
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "error": {
            "type": ["string", "null"]
          }
        }
      },
...
      "LabelRef": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "color": {
            "type": ["string", "null"]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "LinearListTeamsRequest": {
        "type": "object",
        "required": ["connection_id"],
        "properties": {
          "connection_id": {
            "type": "string"
          },
          "cursor": {
            "type": ["string", "null"]
          },
          "limit": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "LinearListTicketsRequest": {
        "type": "object",
        "required": ["connection_id", "team_id"],
        "properties": {
          "connection_id": {
            "type": "string"
          },
          "cursor": {
            "type": ["string", "null"]
          },
          "limit": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 0
          },
          "query": {
            "type": ["string", "null"]
          },
          "team_id": {
            "type": "string"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "content": {
            "type": ["string", "null"]
          },
          "createdAt": {
            "type": ["string", "null"]
          },
          "id": {
            "type": "string"
          },
          "messageType": {
            "type": ["string", "null"]
          }
        }
      },
...
      "PersonRef": {
        "type": "object",
        "properties": {
          "avatar_url": {
            "type": ["string", "null"]
          },
          "email": {
            "type": ["string", "null"]
          },
          "id": {
            "type": ["string", "null"]
          },
          "name": {
            "type": ["string", "null"]
          }
        }
      },
...
      "PullRequestDetail": {
        "type": "object",
        "required": ["is_draft", "is_merged"],
        "properties": {
          "is_draft": {
            "type": "boolean"
          },
          "is_merged": {
            "type": "boolean"
          },
          "merged_at": {
            "type": ["string", "null"]
          },
          "merged_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PersonRef"
              }
            ]
          },
          "source_branch": {
            "type": ["string", "null"]
          },
          "target_branch": {
            "type": ["string", "null"]
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": {
            "type": "string"
          },
          "messageType": {
            "type": "string"
          },
          "sourceId": {
            "type": ["string", "null"]
          }
        }
      },
...
      "TicketKind": {
        "type": "string",
        "enum": ["issue", "pull_request"]
      },
      "TicketPage": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketSummary"
            }
          },
          "next_cursor": {
            "type": ["string", "null"]
          }
        }
      },
      "TicketPriority": {
        "type": "string",
        "enum": ["urgent", "high", "medium", "low", "none"]
      },
      "TicketProviderType": {
        "type": "string",
        "enum": ["github", "linear"]
      },
      "TicketState": {
        "type": "string",
        "enum": ["backlog", "open", "in_progress", "done", "closed"]
      },
      "TicketSummary": {
        "type": "object",
        "required": [
          "provider",
          "kind",
          "id",
          "collection",
          "title",
          "state",
          "assignees",
          "labels",
          "url",
          "created_at",
          "updated_at",
          "raw"
        ],
        "properties": {
          "assignees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonRef"
            }
          },
          "author": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PersonRef"
              }
            ]
          },
          "closed_at": {
            "type": ["string", "null"],
            "description": "ISO 8601."
          },
          "collection": {
            "$ref": "#/components/schemas/CollectionRef"
          },
          "created_at": {
            "type": "string",
            "description": "ISO 8601."
          },
          "id": {
            "type": "string",
            "description": "Provider-specific unique ID."
          },
          "kind": {
            "$ref": "#/components/schemas/TicketKind"
          },
          "labels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabelRef"
            }
          },
          "number": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "Numeric identifier (GitHub: number, Linear: number).",
            "minimum": 0
          },
          "priority": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "provider": {
            "$ref": "#/components/schemas/TicketProviderType"
          },
          "pull_request": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PullRequestDetail",
                "description": "Present only when kind == PullRequest."
              }
            ]
          },
          "raw": {
            "type": "string",
            "description": "Raw provider JSON for forward compatibility."
          },
          "state": {
            "$ref": "#/components/schemas/TicketState"
          },
          "state_detail": {
            "type": ["string", "null"],
            "description": "Provider's original state string for display (e.g. \"In Review\", \"merged\")."
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "description": "ISO 8601."
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    }
  },
...
}